lib/sign: add bucket_name arg to s3 req signer
This allows the signing code to easily determine the bucket resource.
Until now, the URL had been checked against the static hostname
s3.amazonaws.com, to determine if and where the bucket string was
present. However, this behaviour is not possible for redirected
requests, which take a URL host of s3-XXXXX.amazonaws.com.