cli: initialise create and ls args at parse time
When entering commands at the elasto> prompt, the cli_args structure is
reused for successive commands.
As a result, any fields used in the cmd handle() callback must be
explicitly initialised in parse(), otherwise values from a prior command
may be used.

This change fixes a bug where an S3 ls (without args) after a create
attempted to list the bucket specified in the prior create command.