lib/xml: handle finder everytime the query is matched
Until now, a finder matching the current path was handled and then moved
to the found list, with the exception of the path callback finder type.

This change removes the found list, so all finders now have the same
behaviour as the path callback finder type. If the query is matched
multiple times, then the value will be stashed or the callback will be
executed multiple times.

Signed-off-by: David Disseldorp <ddiss@suse.de>