lib/azure_ssl: store certificate in private tmp dir
Currently, after parsing an Azure publish-settings file, the resulting
SSL certificate is written to local disk in the same directory as the
publish-settings file using the default umask.
This is behaviour insecure, as the directory may be accessible to more
than just the local user, and the certificate may be readable if the
umask doesn't restrict access to only

This change ensures that the SSL certificate is written to a temporary
directory that is only accessible to the creator.

Signed-off-by: David Disseldorp <ddiss@suse.de>