lib/file: add support for directory opens
This change adds a new ELASTO_FOPEN_DIRECTORY flag for elasto_fopen(),
which then returns a file descriptor representing a directory object.

With the Azure Page Blob file backend, this flag allows for the opening
of container, account and root paths.

Further changes are needed to process/block directory handles for IO and
other operations. Readdir support will also be added in future.

Signed-off-by: David Disseldorp <ddiss@suse.de>