data: add ELASTO_DATA_CB callback buffer type
This allows IO callers to handle the filling and emptying of data
buffers in an on-demand fashion.

Ownership of a data-in buffer transfers from the connection layer to the
requester layer via in_cb invocation. After which the requester layer is
responsible for cleanup via free().

Vice-versa for data-out buffers: on out_cb invocation, the requester
layer transfers ownership of the buffer to the connection layer, which
will free the buffer after use.

data-in buffers are currently implemented using a memory copy of
incoming data. data-out buffers are optimised to avoid memory copies, by
utilising libevent's nice evbuffer_add_reference() hook.

Signed-off-by: David Disseldorp <ddiss@suse.de>