lib/azure_blob_req: num_blks param for Put Block List
The Azure Put Block List request requires a buffer to hold XML
describing each uploaded portion of a multi-part upload.
Until now, the buffer has been a fixed size (2K), which is too small for
large multi-part uploads.

This change adds a num_blks parameter to the request function, so that
the buffer size can be dynamically calculated for allocation.

Signed-off-by: David Disseldorp <ddiss@suse.de>