lib/s3_req: num_parts param for Complete MP Upload
The Azure Complete Multipart Upload request requires a buffer to hold
XML describing each uploaded portion of a multi-part upload.
Until now, the buffer has been a fixed size (2K).

This change adds a num_blks parameter to the request function, so that
the buffer size can be dynamically calculated for allocation.

Signed-off-by: David Disseldorp <ddiss@suse.de>