lib/file/afs: handle mgmt and IO connections separately
Azure requests dealing with account manipulation are generally sent to
the management host, as opposed to share, file and directory IO
requests, which are normally sent to an account specific host.

This change replaces the Azure File Service connection structure with
two conn structs nested under the private afs_fh structure: mgmt_conn
and io_conn.

In future this will allow for the removal of per-request host
reconnects.

Signed-off-by: David Disseldorp <ddiss@suse.de>