lib/conn: server connection Keep-Alive support
Finally remove the suboptimal reconnect-per-request logic currently used
for request dispatch.
Connection initialisation now takes a host parameter, and establishes a
persistent connection to the server through the use of the Connection:
Keep-Alive header.
If the connection is dropped for whatever reason, then reestablishment
occurs prior to subsequent request dispatch.

Bug: https://github.com/elastocloud/elasto/issues/11

Signed-off-by: David Disseldorp <ddiss@suse.de>