lib/conn: use struct event to track entire op exchange
An op txrx can result in multiple request/response exchanges, in the
case of retry or redirect. This change adds a struct event to struct
conn_op, which flags completion only when the entire req/rsp exchange
for the op is finished.

This allows for upper layers to dispatch and track multiple outstanding
requests, through the use of a new elasto_conn_op_tx() function.

Signed-off-by: David Disseldorp <ddiss@suse.de>