cli: build command source with cli binary
c8c46e022f2ae1e67d3df654661e09a090b1cf10 split the command source out as
a separate static library, which resulted in the constructors (which
register the command) not being called.

Move command source back in with the client, but keep the utils (which
are shared with cm_unity) in a separate static library.

Bug: https://github.com/elastocloud/elasto/issues/25

Signed-off-by: David Disseldorp <ddiss@suse.de>